% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_simulation.R
\name{create_state_df}
\alias{create_state_df}
\title{Creates a \code{state_df} object}
\usage{
create_state_df(
  n_cases,
  sim.params,
  sim.status,
  initialize = FALSE,
  import = FALSE
)
}
\arguments{
\item{n_cases}{The number of cases to initialize in this dataframe}

\item{sim.params}{The \code{sim.params} object containing disease and simulation parameters}

\item{sim.status}{The \code{sim.status} state vector}

\item{initialize}{A boolean indicating whether these cases are the first cases of the
simulation. Defaults to \code{FALSE}.}

\item{import}{A boolean indicating whether these new cases are imported cases instead
of secondary infections. Defaults to \code{FALSE}.}
}
\value{
A \code{state_df} dataframe object
}
\description{
This function creates a \code{state_df} dataframe object, which is used to track
active cases in the simulation.
}
\details{
The simulation's main \code{state_df} dataframe is created with this
function at the beginning of the simulation. When new cases are added (secondary or
imported infections), a \code{state_df} dataframe is created for these new cases
by \code{\link{generate_secondary_infections}} and \code{\link{generate_imported_infections}}.
The new \code{state_df} object is merged with the main \code{state_df} within
\code{\link{step_simulation}}.

The columns of this dataframe are: \code{case_id}, \code{status}, \code{is_traced},
\code{is_symptomatic}, \code{days_infected}, \code{incubation_length},
\code{isolation_delay}, \code{infection_length}, \code{sd_factor},
\code{n_sec_infects}, and \code{serial_intervals}. The meaning of these columns
will be documented elsewhere. Most of the columns are constant and are initialized
in this function via calls to related functions in draw_distributions.R. All time lengths
are in days and represent a relative time, not absolute times. The
\code{status}, \code{n_sec_infects} and \code{serial_intervals} columns are updated
over the course of the simulation as cases progress through infection states. The
\code{n_sec_infects} counter and corresponding \code{serial_intervals} list have its
elements removed as the secondary infections are loaded into the simulation. As cases
become isolated or inactive and no longer cause infections, they are removed from this
dataframe.

A companion dataframe, \code{rec_df}, is meant to be a complete record of all cases
that occurred in the simulation. This object is created by \code{\link{create_record_df}}
and tracks all times as the absolute amount of time passed since the simulation start.
}
