% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_contact_rate}
\alias{draw_contact_rate}
\title{Assign physical distancing behaviour to new cases}
\usage{
draw_contact_rate(n_cases, sim_params, sim_status)
}
\arguments{
\item{n_cases}{Number of cases required}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters.
Here, the sim_params$phys_dist_params object contains the information
needed. This list should have the following entries:
\itemize{
\item \code{pd_pop_frac}: A number for the fraction of the general population
that is practicing distancing. Set this to 0 or 1 to have everyone act the same.
\item \code{pd_contact_rate1}: The contact rate of the distancing population before
time \code{pd_change_t}. Set this to 1 to have the distancing population not begin
distancing until a later time.
\item \code{pd_contact_rate2}: The contact rate of the distancing population after
time \code{pd_change_t}. Set this to 1 to have the distancing population stop
distancing at a certain time.
\item \code{pd_change_t}: The simulation time where the distancing population's
contact rate changes. Set this to 0 or a very high number to have only one rate.
}}

\item{sim_status}{\code{sim_status} object (a list) containing simulation state vector}
}
\value{
A vector of length n for the contact_rate (double)
}
\description{
Distancing behaviour is defined as a number between 0 and 1, representing the
relative number of contacts the subject encounters (i.e. 0.6 means the subject
reduces contact to 60\% of the non-distancing population.)
}
\details{
Current implementation assume that either the entire population has the same
distancing behaviour or that the population can be divided into two groups:
one that is practicing distancing and one that is not. User provides parameters such
as the fraction of the general population that is distancing and the relative contact
rate of the distancing population. This relative contact rate can change once, at the
time point provided by the user.

Note: The ratio of distancers to non-distancers of new cases is not the same as the
ratio amongst the general population (user-provided values). This is because new cases
do not arise from the general population but from the population that is in contact with
others. Therefore, the fraction of distancers being infected would generally be lower than
the fraction of distancers in the general population.
}
