% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_generation_interval}
\alias{draw_generation_interval}
\title{Generate generation intervals for potential secondary infections}
\usage{
draw_generation_interval(n, incubation_length, sim_params)
}
\arguments{
\item{n}{Number of potential secondary infections for this case}

\item{incubation_length}{Incubation period of this case (used in skew normal distribution only)}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters, where
all of the information needed to describe the distribution is found within
\code{sim_params$generation_int_params}. Current distributions possible are:
\itemize{
\item \code{skew_norm}: Drawn from skew normal distribution with xi set
to the case's incubation period, and attributes "omega" and "alpha" given
in \code{sim_params$generation_int_params}. This is what Hellewell uses.
\item \code{gamma}: Drawn from a gamma distribution with attributes
"shape" and "rate" given in \code{sim_params$generation_int_params}.
}}
}
\value{
A vector of length n for generation intervals of the case's potential secondary infections
}
\description{
Generation intervals for each case's potential secondary infections are drawn from
distribution defined in \code{sim_params$generation_int_params}
}
