% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_infection_length}
\alias{draw_infection_length}
\title{Determine infection length of new cases}
\usage{
draw_infection_length(n, sim.params)
}
\arguments{
\item{n}{Number of cases required}

\item{sim.params}{\code{sim.params} object (a list) containing simulation parameters, here,
the \code{sim.params$infect.dur} value is used the infection length in days.}
}
\value{
A vector of length n for infection duration (double)
}
\description{
Currently a constant number for all cases
}
