% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_isolation_delay_period}
\alias{draw_isolation_delay_period}
\title{Draw onset-to-isolation periods for new cases from distribution
defined in \code{sim.params$iso_delay_params}}
\usage{
draw_isolation_delay_period(state_df, sim.params)
}
\arguments{
\item{state_df}{\code{state_df} object for the simulation}

\item{sim.params}{\code{sim.params} object (a list) containing simulation parameters, where
all of the information needed to describe the distribution is found within
\code{sim.params$iso_delay_params}. Current distributions possible are:
\itemize{
\item \code{uniform_delay}: Delay is drawn from uniform distribution with
attributes "min" and "max" given in \code{sim.params$iso_delay_params}.
for traced cases. Untraced cases have a further delay of "delay" days.
\item \code{Hellewell}: Delay is drawn from a Weibull distribution with
attributes "shape" and "scale" given in \code{sim.params$iso_delay_params}.
Traced cases have their delay set to zero days.
}}
}
\value{
A vector of length n for case onset-to-isolation period (double)
}
\description{
Draw onset-to-isolation periods for new cases from distribution
defined in \code{sim.params$iso_delay_params}
}
