% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_sd_factor}
\alias{draw_sd_factor}
\title{Assign social distancing behaviour to new cases}
\usage{
draw_sd_factor(
  n_cases,
  sim.params,
  sim.status,
  initialize = FALSE,
  import = FALSE
)
}
\arguments{
\item{n_cases}{Number of cases required}

\item{sim.params}{\code{sim.params} object (a list) containing simulation parameters.
Here, the sim.params$social_dist_params object contains the information
needed. This list should have the following entries:
\itemize{
\item \code{sd_factors}: A list of social distancing factors with length
equal to the number of different social distancing groups.
\item \code{p.group}: A list of probabilities of being in one of the social
distancing groups. Length should be the same as \code{sd_factors}.
\item \code{delay}: The simulation day number where social distancing
begins. All new cases prior to this time will have no distancing factor
applied (i.e. factor is set to 1).
}}

\item{sim.status}{\code{sim.status} object (a list) containing simulation state vector}

\item{initialize}{A boolean indicating whether these cases are the first cases of the
simulation (sets sd_factor to 1). Defaults to \code{FALSE}.}

\item{import}{A boolean indicating whether these new cases are imported cases instead
of secondary infections. Imported cases do not practice social distancing
(sd_factor set to 1). Defaults to \code{FALSE}.}
}
\value{
A vector of length n for the sd_factor (double)
}
\description{
Input parameters determine the social distancing behaviour for different population groups
as well as the probability of a new case being in a given group. Distancing behaviour is
defined as a number between 0 and 1, representing the relative number of contacts the
subject encounters after distancing begins (i.e. 0.6 means the subject reduces contact
to 60\% of the normal amount.) There is also a parameter to determine the simulation day
number where this behaviour begins.
}
