% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_serial_interval}
\alias{draw_serial_interval}
\title{Draw serial intervals for each case's potential secondary infections from distribution
defined in \code{sim.params$serial_int_params}}
\usage{
draw_serial_interval(n, incubation_length, sim.params)
}
\arguments{
\item{n}{Number of potential secondary infections for this case}

\item{incubation_length}{Incubation period of this case (used in skewed normal distribution only)}

\item{sim.params}{\code{sim.params} object (a list) containing simulation parameters, where
all of the information needed to describe the distribution is found within
\code{sim.params$serial_int_params}. Current distributions possible are:
\itemize{
\item \code{dummy}: Uniform number between 1 and 5 (used for testing only)
\item \code{skewed_norm}: Drawn from skewed normal distribution with xi set
to the case's incubation period, and attributes "omega" and "alpha" given
in \code{sim.params$serial_int_params}. This is what Hellewell uses.
\item \code{gamma}: Drawn from a gamma distribution with attributes
"shape" and "rate" given in \code{sim.params$serial_int_params}.
}}
}
\value{
A vector of length n for serial intervals of the case's potential secondary infections
}
\description{
Draw serial intervals for each case's potential secondary infections from distribution
defined in \code{sim.params$serial_int_params}
}
