% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_trace_app_compliance}
\alias{draw_trace_app_compliance}
\title{Determine whether a case will comply with tracing app instructions}
\usage{
draw_trace_app_compliance(state_df, sim_params)
}
\arguments{
\item{state_df}{\code{state_df} object for the new cases}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters, here,
the \code{sim_params$p_trace_app_comp} value is used as the probability of
contact tracing app compliance.}
}
\value{
A boolean vector of length n for each case's app compliance status
}
\description{
Draws a uniform random number and compares to app compliance probability. This status
is determined (but never used) even for cases that aren't traced by the app.
}
