% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_trace_app_user_status}
\alias{draw_trace_app_user_status}
\title{Determine whether a case is using a contact tracing app}
\usage{
draw_trace_app_user_status(n, sim_params)
}
\arguments{
\item{n}{Number of cases to consider}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters, here,
the \code{sim_params$p_trace_app} value is used as the probability of
contact tracing app usage.}
}
\value{
A boolean vector of length n for each case's app usage status
}
\description{
Draws a uniform random number and compares to app usage probability to determine
whether a case is an app user. For a case to be traced via the app, the
secondary case and the primary/index case must both be app users.
}
