% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distributions.R
\name{draw_traced_by_app}
\alias{draw_traced_by_app}
\title{Helper function to check primary and secondary trace_app_user status}
\usage{
draw_traced_by_app(state_df, primary_state_df, primary_case_ids)
}
\arguments{
\item{state_df}{\code{state_df} object for the new cases}

\item{primary_state_df}{The \code{state_df} object for the index/primary cases.}

\item{primary_case_ids}{A list of case_ids for the primary cases.}
}
\value{
A boolean vector of length n for each case's app tracing status
}
\description{
For a case to be traced via the app, the secondary case and the
primary/index case must both be app users.
}
