% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.branch.sim-package.R
\docType{package}
\name{epi.branch.sim-package}
\alias{epi.branch.sim}
\alias{epi.branch.sim-package}
\title{epi.branch.sim: Simulates an Epidemic Outbreak with a Branching Process}
\description{
Simulates an epidemic outbreak with a branching process from a number of 
    initial seed cases. Every simulated case has disease and intervention parameters 
    stochastically drawn from user-provided distributions. These parameters determine 
    which secondary cases occur, which in turn create the next generation of cases. 
    This simulation implements modelling of interventions such as manual/conventional 
    contact tracing, app-based contact tracing, quarantine, self-isolation, and 
    reduction of contacts through physical distancing. This simulation also includes 
    the option of adding new cases via importation, which may have different 
    intervention policies. The simulation itself outputs a dataframe that summarizes 
    every case generated during the course of the simulation. This can be used to 
    generate metrics or other statistics across a large number of runs. Some example 
    driver functions to do this are provided.
}
\references{

}
\author{
\strong{Maintainer}: Henry Ngo \email{Henry.Ngo@gov.bc.ca}

Authors:
\itemize{
  \item Rebeca Cardim Falcão
  \item Michael Otterstatter
}

Other contributors:
\itemize{
  \item Michael Irvine [contributor]
  \item Ioana Sevcenco [contributor]
  \item Caroline Colijn [contributor]
  \item Daniel Coombs [contributor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
