% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_infections.R
\name{generate_imported_infections}
\alias{generate_imported_infections}
\title{Adds imported infections (new cases) to simulation}
\usage{
generate_imported_infections(sim.params, sim.status)
}
\arguments{
\item{sim.params}{\code{sim.params} object (a list) containing simulation parameters}

\item{sim.status}{\code{sim.status} object (a list) containing simulation state vector}
}
\value{
A list containing
\itemize{
\item \code{n_imported_infections} - Integer with number of imported cases at this time step.
\item \code{import_source} - String describing the source of these new cases. This information
will be recorded in \code{state_df} and \code{record_df}. For now, this will always be "imported".
}
}
\description{
The daily risk of imported infections is defined from an external data file which
contains the mean number of infected cases arriving per day. This risk level is
loaded into the simulation as \code{sim.params$import_params}. This function "loads"
the infections into the simulation at integer time steps, drawing a number of new
infections as Poisson with mean equal to the daily risk.
}
