% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_new_infections.R
\name{generate_secondary_infections}
\alias{generate_secondary_infections}
\title{Adds secondary infections to simulation}
\usage{
generate_secondary_infections(state_df, sim_params)
}
\arguments{
\item{state_df}{\code{state_df} object for the simulation}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters}
}
\value{
A list containing
\itemize{
\item \code{state_df} - The updated \code{state_df} object for the simulation
\item \code{sec_infection_sources} - A vector of \code{case_id}s causing the infections.
If a case causes multiple secondary infections, the ID is repeated. For example,
c(1,5,5,7) means case 1 caused 1 infection, case 5 caused 2 infections and case 7
caused 1 infection. If no new secondary infections are generated in this time step,
this value is \code{NULL}.
}
}
\description{
Potential secondary infections are generated at the time the primary
case was generated. This function "loads" the infections into the simulation
at the correct time. It also modifies the \code{state_df} to account for the
loaded infection.
}
