% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_simulation.R
\name{initialize_sim_params}
\alias{initialize_sim_params}
\title{Creates \code{sim.params} object for global simulation parameters}
\usage{
initialize_sim_params(
  R0,
  infect.dur,
  vary.trace,
  p.trace,
  cluster_breaks,
  cluster_p.trace,
  p.symp,
  dt,
  incub_params,
  serial_int_params,
  iso_delay_params,
  sec_infect_params,
  import_params,
  social_dist_params
)
}
\arguments{
\item{R0}{Reproductive number of disease. Used in \code{\link{draw_sec_infects_df}} to
determine the number of potential secondary infections
from a primary case.}

\item{infect.dur}{Duration of infection (in days). This constant value determines when
a case is no longer active. It's also the length of the infectious
period, unless the case is isolated. See \code{\link{draw_infection_length}}}

\item{vary.trace}{Boolean to determine whether contact tracing is a constant or variable
based on the number of new cases in a cluster.
See \code{\link{draw_traced_status}}}

\item{p.trace, cluster_breaks, cluster_p.trace}{Parameters determining contact tracing
probability. See \code{\link{draw_traced_status}}}

\item{p.symp}{Probability that a case is symptomatic.
Used in \code{\link{draw_symptomatic_status}}.}

\item{dt}{Size of the simulation time step (in days).}

\item{incub_params}{Parameters for distribution of incubation length.
See \code{\link{draw_incubation_period}}.}

\item{serial_int_params}{Parameters for distribution of serial intervals.
See \code{\link{draw_serial_interval}}.}

\item{iso_delay_params}{Parameters for distribution of delay from symptom onset to isolation.
See \code{\link{draw_isolation_delay_period}}.}

\item{sec_infect_params}{Parameters to model secondary infections.
See \code{\link{draw_sec_infects_df}}}

\item{import_params}{Parameters to model imported infections.
See \code{\link{generate_imported_infections}}}

\item{social_dist_params}{Parameters to model social/physical distancing or
other interventions that reduce the potential to
cause secondary infections.
See \code{\link{draw_sd_factor}}}
}
\value{
A list containing all of the above entries
}
\description{
This object is a list containing global simulation parameters such as the disease
parameters, either directly defined or through a distribution with user-provided
values. This object also contains some parameters that control the simulation itself.
Many of these values are used in other functions, the major ones are linked here.
}
