% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_simulation.R
\name{initialize_sim_status}
\alias{initialize_sim_status}
\title{Creates \code{sim_status} object, the simulation state vector}
\usage{
initialize_sim_status(start_time, start_n_cases)
}
\arguments{
\item{start_time}{The start time (in days) for the simulation}

\item{start_n_cases}{The number of cases at the start of the simulation.}
}
\value{
A list containing the simulation state vector quantities, which includes:
\itemize{
\item \code{t} Current time (days)
\item \code{last_case_id} The last case_id generated (integer)
}
}
\description{
This object is a list containing the simulation state vector. This contains values that
change with every time step. Note that \code{sim_params} is another important object
containing quantities that remain constant throughout the simulation. Currently, the
only two items in \code{sim_status} are the time and the case_id of the last case
infected (used to generate the next case_id). This function creates the \code{sim_status}
object at the beginning of the simulation and only needs to be called once.
}
