% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{run_scenario_Hellewell}
\alias{run_scenario_Hellewell}
\title{Run scenarios as in Hellewell et al. (2020) paper}
\usage{
run_scenario_Hellewell(
  nsims,
  p_trace,
  R0,
  infections_before_symptoms = "medium",
  iso_delay_length = "short",
  p_symp = 1,
  initial_cases = 20,
  max_cases = 5000
)
}
\arguments{
\item{nsims}{Number of simulations to run in this scenario}

\item{p_trace}{Fraction of new infections that are traced}

\item{R0}{Basic reproduction number for disease}

\item{infections_before_symptoms}{One of "low", "medium" or "high" corresponding to one
of three generation interval distributions considered by
Hellewell et al. These distributions result in <1\%,
15\%, and 30\% of generation interval lengths occuring prior
to symptom onset. "Medium" (15\%) is the default value
and the benchmark used in Hellewell et al.}

\item{iso_delay_length}{One of "short" or "long" corresponding to two distributions used
by Hellewell for the delay from symptom onset to isolation for
untraced symptomatic cases. "Short" is the default and
benchmark value.}

\item{p_symp}{Fraction of cases that are symptomatic. 1 is the default and benchmark value.}

\item{initial_cases}{Number of cases at the start. 20 is the default and benchmark value}

\item{max_cases}{Number of total cases before simulation ends with an "epidemic"
outcome. Defaults to 5000, used by Hellewell et al.}
}
\value{
A named vector with the following entries describing the scenario and
the outcome (percentage of simulations that went extinct and absolute number
of each type of outcome). Namely the entries are: nsims, R0, p_trace,
percentage, nsims_extinct, nsims_epidemic, nsims_ongoing.
}
\description{
For a given tracing fraction and R0 (a "scenario"), count the fraction of simulations
that are controlled. Designed to reproduce results from the source paper for this model: \href{https://doi.org/10.1016/S2214-109X(20)30074-7}{Hellewell et al. "Feasibility of controlling COVID-19 outbreaks by isolation of cases and contacts." \emph{The Lancet Global Health} 2020; 8: E488-E496. DOI: 10.1016/S2214-109X(20)30074-7}
}
\details{
The intention of this driver function is to reproduce Figure 3A from Hellewell et al. which
shows the fraction of outbreaks controlled as a function of R0 and tracing fraction. These
parameters are the positional arguments to this function. The remaining named arguments allow
for consideration of other scenarios examined by Hellewell et al. The named arguments' default
values are Hellewell et al.'s benchmark values but can be changed to reproduce results from
Figure 4 of the source paper.

The secondary intention of this driver function is to provide a simple example of how to
call the exported functions of this module in order to set up simulations objects, run
a number of simulations and return summary statistics/metrics for further analysis.
}
