% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_simulation.R
\name{step_simulation}
\alias{step_simulation}
\title{Takes the simulation one dt timestep forward}
\usage{
step_simulation(sim_status, state_df, rec_df, sim_params)
}
\arguments{
\item{sim_status}{\code{sim_status} list containing simulation state vector}

\item{state_df}{\code{state_df} object for the simulation}

\item{rec_df}{\code{record_df} object for the simulation}

\item{sim_params}{\code{sim_params} object (a list) containing simulation parameters}
}
\value{
A list containing
\itemize{
\item \code{status}   - The updated \code{sim_status} object
\item \code{state}    - The updated \code{state_df} object
\item \code{record}   - The updated \code{record_df} object
\item \code{new_sec_cases} - Number of new secondary infections generated in this time step
\item \code{new_imp_cases} - Number of new imported infections generated in this time step
}
}
\description{
Use the outputs of this function as input for the next step.
}
