% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_simulation.R
\name{create_record_df}
\alias{create_record_df}
\title{Creates a \code{rec_df} object}
\usage{
create_record_df(
  state_df,
  sim_status,
  initialize = FALSE,
  infection_source = NULL
)
}
\arguments{
\item{state_df}{\code{state_df} object for the cases to be generated}

\item{sim_status}{The \code{sim_status} state vector}

\item{initialize}{A boolean indicating whether these cases are the first cases of the
simulation. Defaults to \code{FALSE}.}

\item{infection_source}{A string or vector of case_ids representing the source of each
case's infection. Defaults to \code{NULL}, however, this
parameter cannot be NULL if \code{initialize} is \code{FALSE}.}
}
\value{
A \code{rec_df} dataframe object
}
\description{
This function creates a \code{rec_df} dataframe object, which is used to record all
cases that ever get created within a simulation.
}
\details{
The simulation's main \code{rec_df} dataframe is created with this
function at the beginning of the simulation. When new cases are added (secondary or
imported infections), a \code{rec_df} dataframe is created for these new cases
by \code{\link{generate_secondary_infections}} and \code{\link{generate_imported_infections}}.
The new \code{rec_df} object is merged with the main \code{rec_df} within
\code{\link{step_simulation}}.

The columns of this dataframe are: \code{case_id}, \code{source}, \code{is_traced},
\code{is_trace_app_user}, \code{is_traced_by_app}, \code{is_trace_app_comply},
\code{is_symptomatic}, \code{d_incub}, \code{d_iso_delay}, \code{d_infection},
\code{contact_rate}, \code{n_sec_infects}, \code{d_generation_ints}, \code{t_inf},
\code{t_symp}, \code{t_iso}, \code{t_inact}, \code{cases_inf}, \code{s_status},
and \code{non_infect_generations}. The meaning of these columns will be documented elsewhere.
Columns \code{case_id} through \code{t_inf} as well as column
\code{non_infect_generations} are constant and are initialized when this function
is called, using the provided \code{state_df} object. The columns \code{t_symp},
\code{t_iso} and \code{t_inact} are initialized as \code{NA} and replaced with the actual
simulation time where these events happen. If they do not happen (e.g. an asymptomatic
case will not have a \code{t_symp} or \code{t_iso}), then the value remains as \code{NA}.
The column \code{cases_inf} is meant for a list of case_id of all secondary infections
caused by the given case but is not yet implemented. The column \code{s_status} is updated
by \code{\link{step_simulation}} at the appropriate time in the simulation.

Note that all time lengths are given in days and values in columns beginning with
\code{d_} are a period of days since the relevant reference time while columns
beginning with \code{t_} are absolute time, measured in days since the simulation began.
}
