% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_simulation.R
\name{initialize_sim_params}
\alias{initialize_sim_params}
\title{Creates \code{sim_params} object for global simulation parameters}
\usage{
initialize_sim_params(
  R0,
  infect_dur,
  vary_trace,
  p_trace,
  p_trace_app,
  p_trace_app_comp,
  p_symp,
  dt,
  incub_params,
  generation_int_params,
  iso_delay_params,
  sec_infect_params,
  import_params,
  phys_dist_params
)
}
\arguments{
\item{R0}{Reproductive number of disease. Used in \code{\link{draw_sec_infects_df}} to
determine the number of potential secondary infections
from a primary case.}

\item{infect_dur}{Duration of infection (in days). This constant value determines when
a case is no longer active. It's also the length of the infectious
period, unless the case is isolated. See \code{\link{draw_infection_length}}}

\item{vary_trace}{Boolean to determine whether contact tracing is a constant or variable
based on the number of new cases in a cluster.
See \code{\link{draw_traced_status}}}

\item{p_trace}{Probability that a case will be manually traced.}

\item{p_trace_app}{Probability that a case will be a contact tracing app user.}

\item{p_trace_app_comp}{Probability that a case would comply with contact tracing app recommendations.}

\item{p_symp}{Probability that a case is symptomatic.
Used in \code{\link{draw_symptomatic_status}}.}

\item{dt}{Size of the simulation time step (in days).}

\item{incub_params}{Parameters for distribution of incubation length.
See \code{\link{draw_incubation_period}}.}

\item{generation_int_params}{Parameters for distribution of generation intervals.
See \code{\link{draw_generation_interval}}.}

\item{iso_delay_params}{Parameters for distribution of delay to isolation.
See \code{\link{draw_isolation_delay_period}}.}

\item{sec_infect_params}{Parameters to model secondary infections.
See \code{\link{draw_sec_infects_df}}}

\item{import_params}{Parameters to model imported infections.
See \code{\link{generate_imported_infections}}}

\item{phys_dist_params}{Parameters to model physical (social) distancing or
other interventions that reduce the potential to
cause secondary infections.
See \code{\link{draw_contact_rate}}}
}
\value{
A list containing all of the above entries
}
\description{
This object is a list containing global simulation parameters such as the disease
parameters, either directly defined or through a distribution with user-provided
values. This object also contains some parameters that control the simulation itself.
Many of these values are used in other functions, the major ones are linked here.
}
